requirements.txt
django
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2
